refactor(grid-footer): [TinyGridFooter]refactor tinygridfooter with rederless logic
