+++ packages/renderless/types/index.ts
@@ -97,6 +97,7 @@ export * from './form.type'
 export * from './form-item.type'
 export * from './fullscreen.type'
 export * from './grid.type'
+export * from './grid-footer.type'
 export * from './guide.type'
 export * from './hrapprover.type'
 export * from './image.type'
